requirements.txt
RPi.Gpio==0.6.3